Move problem graph creation to MSolver
