+++ libmamba/include/mamba/core/solver.hpp
@@ -21,6 +21,7 @@
 
 #include "mamba/core/package_info.hpp"
 #include "mamba/core/pool.hpp"
+#include "mamba/core/satisfiability_error.hpp"
 
 #include "match_spec.hpp"
 
@@ -82,6 +83,7 @@ namespace mamba
         [[nodiscard]] std::string problems_to_str() const;
         [[nodiscard]] std::vector<std::string> all_problems() const;
         [[nodiscard]] std::vector<MSolverProblem> all_problems_structured() const;
+        [[nodiscard]] ProblemsGraph problems_graph() const;
         [[nodiscard]] std::string all_problems_to_str() const;
         std::ostream& explain_problems(std::ostream& out) const;
         [[nodiscard]] std::string explain_problems() const;
